Move struct definition to a sepereate package
