+++ internal/datasource/storage/s3.go
@@ -4,8 +4,8 @@ import (
 	"context"
 	"os"
 
-	"github.com/aws/aws-sdk-go-v2/service/dynamodb/types"
 	"github.com/aws/aws-sdk-go-v2/service/s3"
+	"github.com/harryzcy/mailbox/internal/types"
 	"github.com/jhillyerd/enmime"
 )
 
@@ -16,8 +16,8 @@ var (
 type GetEmailResult struct {
 	Text        string
 	HTML        string
-	Attachments Files
-	Inlines     Files
+	Attachments types.Files
+	Inlines     types.Files
 }
 
 // S3Storage is an interface that defines required S3 functions
@@ -80,57 +80,11 @@ func (s s3Storage) DeleteEmail(ctx context.Context, api S3DeleteObjectAPI, messa
 	return nil
 }
 
-type File struct {
-	ContentID         string            `json:"contentId"`
-	ContentType       string            `json:"contentType"`
-	ContentTypeParams map[string]string `json:"contentTypeParams"`
-	FileName          string            `json:"filename"`
-}
-
-func (f File) ToAttributeValue() types.AttributeValue {
-	params := make(map[string]types.AttributeValue)
-	for k, v := range f.ContentTypeParams {
-		params[k] = &types.AttributeValueMemberS{
-			Value: v,
-		}
-	}
-
-	return &types.AttributeValueMemberM{
-		Value: map[string]types.AttributeValue{
-			"contentId": &types.AttributeValueMemberS{
-				Value: f.ContentID,
-			},
-			"contentType": &types.AttributeValueMemberS{
-				Value: f.ContentType,
-			},
-			"contentTypeParams": &types.AttributeValueMemberM{
-				Value: params,
-			},
-			"filename": &types.AttributeValueMemberS{
-				Value: f.FileName,
-			},
-		},
-	}
-}
-
-type Files []File
-
-func (fs Files) ToAttributeValue() types.AttributeValue {
-	value := make([]types.AttributeValue, len(fs))
-	for i, f := range fs {
-		value[i] = f.ToAttributeValue()
-	}
-
-	return &types.AttributeValueMemberL{
-		Value: value,
-	}
-}
-
 // parseFiles parses enmime parts into File slice
-func parseFiles(parts []*enmime.Part) Files {
-	files := make([]File, len(parts))
+func parseFiles(parts []*enmime.Part) types.Files {
+	files := make([]types.File, len(parts))
 	for i, part := range parts {
-		files[i] = File{
+		files[i] = types.File{
 			ContentID:         part.ContentID,
 			ContentType:       part.ContentType,
 			ContentTypeParams: part.ContentTypeParams,
